Prep for classed highlighting
Give `CodeBlock` access to the start and end events because classed highlighting needs to adjust the pre element and close spans.
Rename get_highlighter and make it return everything needed for highlighting so that we don't need the config after a codeblock is created.
Prep the code block for classed highlighting which will need to store different state from inline highlighting.